+++ R/wflow_html.R
@@ -303,17 +303,18 @@ wflow_html <- function(...) {
     # Add session information at the end
     if (has_code && wflow_opts$sessioninfo != "") {
       sessioninfo <- glue::glue('
+      <p>
       <button type="button" class="btn btn-default"
-        data-toggle="collapse" data-target="#workflowr-sessioninfo">
+        data-toggle="collapse" data-target="#workflowr-sessioninfo"
+        style = "display: block;">
         Session information
       </button>
+      </p>
 
       <div id="workflowr-sessioninfo" class="collapse">
-      <p>
       ```{{r session-info-chunk-inserted-by-workflowr}}
       {wflow_opts$sessioninfo}
       ```
-      </p>
       </div>
       ')
       # If there is a bibliography, make sure it appears before the session